requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
mysqlclient==2.2.0
python-dotenv==1.0.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
